Début exo 4, exercice 3 complété

+++ Java_Atelier_1/Variables.java
@@ -3,7 +3,7 @@ package Java_Atelier_1;
 public class Variables {
 	public void Sqr_root() {
 		// imprime la racine carrée de a * a + b * b
-		float a = (float) 3.0;
+		float a = 3.0F;
 		double b = 4;
 		float c;
